refactor: 헤더의 유효성 검사는 AuthService가 아닌 AuthInterceptor에서 검사 하도록 수행
헤더 유효성 검사 로직의 분리
