+++ src/main/java/gift/config/WebConfig.java
@@ -1,5 +1,6 @@
 package gift.config;
 
+import gift.controller.auth.AuthInterceptor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
