+++ src/main/java/gift/service/auth/AuthService.java
@@ -53,12 +53,6 @@ public class AuthService {
         return MemberRole.valueOf(role.toString());
     }
 
-    public String getTokenWithAuthorizationHeader(String authorizationHeader) {
-        var header = authorizationHeader.split(" ");
-        if (header.length != 2) throw new IllegalArgumentException("잘못된 헤더 정보입니다.");
-        return header[1];
-    }
-
     private SecretKey getSecretKey() {
         return Keys.hmacShaKeyFor(secretKey.getBytes());
     }
